refactor(templates): drop the v2 suffix from the cover-letter package
Third of the four per-family suffix drops. Move the coverletter.v2
sub-packages (data, components, presets) up to templates.coverletter and
rewrite every import and doc reference across the engine, tests, and
examples. The cv.v2 cross-imports (CvIdentity, CvName, SectionLookup) are
untouched until the cv rename. Pure rename — the visual parity test passes
against the unchanged committed baselines, so every letter preset renders
identically.

Tests: ./mvnw verify javadoc:javadoc -pl . — 1378 tests, 0 failures, javadoc
clean; examples + benchmarks compile; perf-smoke + examples-generation smoke
(85) green.
