+++ examples/src/main/java/com/demcha/examples/GenerateAllExamples.java
@@ -47,21 +47,21 @@ import com.demcha.examples.flagships.FeatureCatalogExample;
 import com.demcha.examples.flagships.FinancialReportExample;
 import com.demcha.examples.flagships.MasterShowcaseExample;
 import com.demcha.examples.flagships.ModuleFirstFileExample;
-import com.demcha.examples.templates.coverletter.v2.CvBlueBannerLetterV2Example;
-import com.demcha.examples.templates.coverletter.v2.CvBoxedSectionsLetterV2Example;
-import com.demcha.examples.templates.coverletter.v2.CvCenteredHeadlineLetterV2Example;
-import com.demcha.examples.templates.coverletter.v2.CvClassicSerifLetterV2Example;
-import com.demcha.examples.templates.coverletter.v2.CvCompactMonoLetterV2Example;
-import com.demcha.examples.templates.coverletter.v2.CvEditorialBlueLetterV2Example;
-import com.demcha.examples.templates.coverletter.v2.CvEngineeringResumeLetterV2Example;
-import com.demcha.examples.templates.coverletter.v2.CvExecutiveLetterV2Example;
-import com.demcha.examples.templates.coverletter.v2.CvMintEditorialLetterV2Example;
-import com.demcha.examples.templates.coverletter.v2.CvModernProfessionalLetterV2Example;
-import com.demcha.examples.templates.coverletter.v2.CvMonogramSidebarLetterV2Example;
-import com.demcha.examples.templates.coverletter.v2.CvNordicCleanLetterV2Example;
-import com.demcha.examples.templates.coverletter.v2.CvPanelLetterV2Example;
-import com.demcha.examples.templates.coverletter.v2.CvSidebarPortraitLetterV2Example;
-import com.demcha.examples.templates.coverletter.v2.CvTimelineMinimalLetterV2Example;
+import com.demcha.examples.templates.coverletter.CvBlueBannerLetterV2Example;
+import com.demcha.examples.templates.coverletter.CvBoxedSectionsLetterV2Example;
+import com.demcha.examples.templates.coverletter.CvCenteredHeadlineLetterV2Example;
+import com.demcha.examples.templates.coverletter.CvClassicSerifLetterV2Example;
+import com.demcha.examples.templates.coverletter.CvCompactMonoLetterV2Example;
+import com.demcha.examples.templates.coverletter.CvEditorialBlueLetterV2Example;
+import com.demcha.examples.templates.coverletter.CvEngineeringResumeLetterV2Example;
+import com.demcha.examples.templates.coverletter.CvExecutiveLetterV2Example;
+import com.demcha.examples.templates.coverletter.CvMintEditorialLetterV2Example;
+import com.demcha.examples.templates.coverletter.CvModernProfessionalLetterV2Example;
+import com.demcha.examples.templates.coverletter.CvMonogramSidebarLetterV2Example;
+import com.demcha.examples.templates.coverletter.CvNordicCleanLetterV2Example;
+import com.demcha.examples.templates.coverletter.CvPanelLetterV2Example;
+import com.demcha.examples.templates.coverletter.CvSidebarPortraitLetterV2Example;
+import com.demcha.examples.templates.coverletter.CvTimelineMinimalLetterV2Example;
 import com.demcha.examples.templates.cv.v2.CvBlueBannerExample;
 import com.demcha.examples.templates.cv.v2.CvBoxedV2Example;
 import com.demcha.examples.templates.cv.v2.CvCenteredHeadlineExample;
