+++ src/main/java/com/demcha/compose/document/templates/coverletter/components/package-info.java
@@ -1,10 +1,10 @@
 /**
  * Shared rendering components for Templates v2 cover letters.
  *
- * <p>{@link com.demcha.compose.document.templates.coverletter.v2.components.LetterBody}
+ * <p>{@link com.demcha.compose.document.templates.coverletter.components.LetterBody}
  * is the letter analog of the CV {@code SectionDispatcher}: every
  * letter preset delegates its greeting / paragraphs / closing to it so
  * all letters share one reading rhythm and inline-markdown handling
  * (via the reused {@code core.text.RichParagraphRenderer}).</p>
  */
-package com.demcha.compose.document.templates.coverletter.v2.components;
+package com.demcha.compose.document.templates.coverletter.components;
