+++ src/main/java/com/demcha/compose/document/templates/coverletter/data/package-info.java
@@ -1,10 +1,10 @@
 /**
  * User-facing data records for Templates v2 cover letters.
  *
- * <p>{@link com.demcha.compose.document.templates.coverletter.v2.data.CoverLetterDocument}
+ * <p>{@link com.demcha.compose.document.templates.coverletter.data.CoverLetterDocument}
  * is the single input type — it reuses
  * {@link com.demcha.compose.document.templates.cv.v2.data.CvIdentity}
  * for the masthead so a CV and its paired letter share one identity
  * object and render identical headers.</p>
  */
-package com.demcha.compose.document.templates.coverletter.v2.data;
+package com.demcha.compose.document.templates.coverletter.data;
