+++ src/main/java/com/demcha/compose/document/templates/coverletter/package-info.java
@@ -1,9 +1,9 @@
 /**
  * Cover-letter template family.
  *
- * <p>The shipping cover-letter surface is the layered stack under
- * {@code coverletter.v2} (data / theme / components / widgets / presets),
- * built on the shared {@code templates.core} layer and a {@code BrandTheme}.</p>
+ * <p>The cover-letter surface is a layered stack (data / components /
+ * presets) built on the shared {@code templates.core} layer and a
+ * {@code BrandTheme}.</p>
  *
  * <p><strong>Naming note:</strong> the user-facing concept is "cover-letter"
  * with a hyphen, but Java packages cannot contain hyphens. The package name
